feat: Add new Navigation folders and restructure layouts and add side navigation
